redesigning the entire application
