+++ src/main/java/UsageExample.java
@@ -1,0 +1,10 @@
+import util.engine.Markovener;
+
+import java.io.IOException;
+import java.nio.file.Path;
+
+public class UsageExample {
+   public static void main(String[] args) throws IOException {
+      Markovener.createByNgrams("Markovener is cool!", 2).save(Path.of("src","test","outputtest"));
+   }
+}
