--- src/test/java/enginetest/MapperTest.java
@@ -1,5 +1,0 @@
-package enginetest;
-
-public class MapperTest {
-   /* TODO */
-}
