+++ src/test/outputtest/output.json
@@ -1,0 +1,1 @@
+{"oo":["l!"]," c":["oo"],"ko":["ve"],"ov":["en"]," i":["s "],"en":["er"],"is":[" c"],"co":["ol"],"s ":["co"],"er":[" i"],"r ":["is"],"ve":["ne"],"ar":["ko"],"Ma":["rk"],"ne":["r "],"rk":["ov"]}
